pyproject.toml
[project]
name = "ng-datasets-be"
version = "0.1.0"
description = "Backend API for Nigerian Datasets project"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dj-database-url>=2.2.0",
    "django>=5.1.0",
    "django-background-tasks>=1.2.8",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.28.0",
    "fastparquet>=2024.11.0",
    "google-cloud-storage>=3.3.1",
    "gunicorn>=23.0.0",
    "markdown>=3.8.2",
    "numpy>=1.24.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pandas-stubs>=2.3.2.250827",
    "psycopg2-binary>=2.9.10",
    "pyarrow>=21.0.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.3.0",
    "sentence-transformers>=2.2.2",
    "whitenoise>=6.8.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "django-stubs>=5.2.2",
    "djangorestframework-stubs>=3.16.2",
    "isort>=6.0.1",
    "mypy>=1.17.1",
    "pre-commit>=4.3.0",
    "ruff>=0.12.11",
    "types-pyyaml>=6.0.12.20250822",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "DJ"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["mypy_django_plugin.main"]

[[tool.mypy.overrides]]
module = ["google.*", "background_task.*"]
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "backend.settings"
